[Remove] elmina2 file
