+++ poke_shiny.py
@@ -7,9 +7,39 @@ tkinterを使ってるのでpython3専用。
 しかし作るのは楽しい。
 '''
 
-import elmina2
-import tkinter as Tk
 import os
+import random
+import tkinter as Tk
+
+
+# 擬似色違い孵化
+# 0通常出現率1/4096 1ひかおま所持3/same 2国際孵化6/same 3ひかおま国際孵化8/same
+def poke_shiny(x):
+    success = times = 0
+    while success == 0:
+        a = random.randint(1, 4096)
+        times += 1
+        if x == 0:
+            if a == 1:
+                success = 1
+            else:
+                continue
+        elif x == 1:
+            if a <= 3:
+                success = 1
+            else:
+                continue
+        elif x == 2:
+            if a <= 6:
+                success = 1
+            else:
+                continue
+        else:
+            if a <= 8:
+                success = 1
+            else:
+                continue
+    return '\n\nHey! I\'m %s th brother.' % times
 
 
 class TkRoot(Tk.Frame):
@@ -38,7 +68,7 @@ class TkRoot(Tk.Frame):
         if_second.set(0)
 
         def foo():
-            result.set(elmina2.poke_shiny(m.get()))
+            result.set(poke_shiny(m.get()))
             if if_second.get() < 1:
                 Tk.Label(self, image=self.img).pack()
             if_second.set(1)
